add python and c#

+++ reset.py
@@ -8,6 +8,7 @@ Usage: `./reset.py`
 """
 
 from lib import (
+    check_image_exists,
     delete_containers_by_image, 
     delete_image,
     LANGUAGES
@@ -17,8 +18,9 @@ def reset():
     images = [lang + "_fork_bomb" for lang in LANGUAGES]
 
     for img in images:
-        delete_containers_by_image(img)
-        delete_image(img)
+        if check_image_exists(img):
+            delete_containers_by_image(img)
+            delete_image(img)
 
 
 if __name__ == "__main__": reset()
